+++ R/analyze.lmerModLmerTest.R
@@ -36,8 +36,9 @@ analyze.lmerModLmerTest <- function(x, CI=95, effsize_rules="cohen1988", ...) {
   outcome <- predictors[[1]]
   predictors <- tail(predictors, -1)
 
-  R2m <- MuMIn::r.squaredGLMM(fit)["R2m"]
-  R2c <- MuMIn::r.squaredGLMM(fit)["R2c"]
+  R2 <- MuMIn::r.squaredGLMM(fit)
+  R2m <- R2["R2m"]
+  R2c <- R2["R2c"]
 
 
   # TODO: Bootstrapped p values
@@ -52,36 +53,36 @@ analyze.lmerModLmerTest <- function(x, CI=95, effsize_rules="cohen1988", ...) {
 
   # Summary
   # -------------
-  fitsum <- data.frame(summary(fit)$coefficients)
+  summary <- data.frame(summary(fit)$coefficients)
 
-  fitsum$Variable <- rownames(fitsum)
-  fitsum$Coef <- fitsum$Estimate
-  fitsum$SE <- fitsum$`Std..Error`
-  fitsum$df <- as.numeric(fitsum$df)
-  fitsum$t <- fitsum$`t.value`
-  fitsum$p <- fitsum$`Pr...t..`
+  summary$Variable <- rownames(summary)
+  summary$Coef <- summary$Estimate
+  summary$SE <- summary$`Std..Error`
+  summary$df <- as.numeric(summary$df)
+  summary$t <- summary$`t.value`
+  summary$p <- summary$`Pr...t..`
 
   # standardized coefficients
   stdz <- as.data.frame(MuMIn::std.coef(fit, partial.sd = FALSE))
-  fitsum$Coef.std <- stdz$Estimate
-  fitsum$SE.std <- stdz$`Std. Error`
-  fitsum$Effect_Size <- interpret_d(fitsum$Coef.std, rules = effsize_rules)
+  summary$Coef.std <- stdz$Estimate
+  summary$SE.std <- stdz$`Std. Error`
+  summary$Effect_Size <- c(NA, interpret_d(tail(summary$Coef.std, -1), rules = effsize_rules))
 
-  fitsum <- dplyr::select_(
-    fitsum, "Variable", "Coef", "SE", "t", "df", "Coef.std", "SE.std",
+  summary <- dplyr::select_(
+    summary, "Variable", "Coef", "SE", "t", "df", "Coef.std", "SE.std",
     "p", "Effect_Size"
   )
 
   if (!is.null(CI)) {
     CI_values <- suppressMessages(confint(fit, level = CI / 100))
-    CI_values <- tail(CI_values, n = length(rownames(fitsum)))
-    fitsum$CI_lower <- CI_values[, 1]
-    fitsum$CI_higher <- CI_values[, 2]
+    CI_values <- tail(CI_values, n = length(rownames(summary)))
+    summary$CI_lower <- CI_values[, 1]
+    summary$CI_higher <- CI_values[, 2]
   }
 
 
   # Varnames
-  varnames <- rownames(fitsum)
+  varnames <- rownames(summary)
 
 
 
@@ -95,7 +96,7 @@ analyze.lmerModLmerTest <- function(x, CI=95, effsize_rules="cohen1988", ...) {
 
   # Loop over all variables
   for (varname in varnames) {
-    if (fitsum[varname, "p"] < .1) {
+    if (summary[varname, "p"] < .1) {
       significance <- " "
     } else {
       significance <- " not "
@@ -105,9 +106,9 @@ analyze.lmerModLmerTest <- function(x, CI=95, effsize_rules="cohen1988", ...) {
       CI_text <- paste0(
         ", ",
         CI, "% CI [",
-        format_digit(fitsum[varname, "CI_lower"], null_treshold = 0.0001),
+        format_digit(summary[varname, "CI_lower"], null_treshold = 0.0001),
         ", ",
-        format_digit(fitsum[varname, "CI_higher"], null_treshold = 0.0001),
+        format_digit(summary[varname, "CI_higher"], null_treshold = 0.0001),
         "]"
       )
     } else {
@@ -116,49 +117,55 @@ analyze.lmerModLmerTest <- function(x, CI=95, effsize_rules="cohen1988", ...) {
 
 
 
-    text <- paste0(
-      "The effect of ",
-      varname,
-      " is",
-      significance,
-      "significant (beta = ",
-      format_digit(fitsum[varname, "Coef"], 2), ", SE = ",
-      format_digit(fitsum[varname, "SE"], 2),
-      CI_text,
-      ", t(",
-      format_digit(fitsum[varname, "df"], 2), ") = ",
-      format_digit(fitsum[varname, "t"], 2), ", p ",
-      format_p(fitsum[varname, "p"]),
-      ") and can be considered as ",
-      tolower(fitsum[varname, "Effect_Size"]),
-      " (std. beta = ",
-      format_digit(fitsum[varname, "Coef.std"], 2),
-      ", std. SE = ",
-      format_digit(fitsum[varname, "SE.std"], 2), ")."
-    )
 
     if (varname == "(Intercept)") {
       text <- paste0(
         "The model's intercept is at ",
-        format_digit(fitsum[varname, "Coef"], 2),
+        format_digit(summary[varname, "Coef"], 2),
         " (SE = ",
-        format_digit(fitsum[varname, "SE"], 2),
+        format_digit(summary[varname, "SE"], 2),
         CI_text,
         "). Within this model:"
+      )
+    } else{
+
+      text <- paste0(
+        "The effect of ",
+        varname,
+        " is",
+        significance,
+        "significant (beta = ",
+        format_digit(summary[varname, "Coef"], 2),
+        ", SE = ",
+        format_digit(summary[varname, "SE"], 2),
+        CI_text,
+        ", t(",
+        format_digit(summary[varname, "df"], 2),
+        ") = ",
+        format_digit(summary[varname, "t"], 2),
+        ", p ",
+        format_p(summary[varname, "p"]),
+        ") and can be considered as ",
+        tolower(summary[varname, "Effect_Size"]),
+        " (std. beta = ",
+        format_digit(summary[varname, "Coef.std"], 2),
+        ", std. SE = ",
+        format_digit(summary[varname, "SE.std"], 2),
+        ")."
       )
     }
 
     values$effects[[varname]] <- list(
-      Coef = fitsum[varname, "Coef"],
-      SE = fitsum[varname, "SE"],
-      CI_lower = fitsum[varname, "CI_lower"],
-      CI_higher = fitsum[varname, "CI_higher"],
-      t = fitsum[varname, "t"],
-      df = fitsum[varname, "df"],
-      Coef.std = fitsum[varname, "Coef.std"],
-      SE.std = fitsum[varname, "SE.std"],
-      p = fitsum[varname, "p"],
-      Effect_Size = fitsum[varname, "Effect_Size"],
+      Coef = summary[varname, "Coef"],
+      SE = summary[varname, "SE"],
+      CI_lower = summary[varname, "CI_lower"],
+      CI_higher = summary[varname, "CI_higher"],
+      t = summary[varname, "t"],
+      df = summary[varname, "df"],
+      Coef.std = summary[varname, "Coef.std"],
+      SE.std = summary[varname, "SE.std"],
+      p = summary[varname, "p"],
+      Effect_Size = summary[varname, "Effect_Size"],
       Text = text
     )
   }
@@ -172,8 +179,7 @@ analyze.lmerModLmerTest <- function(x, CI=95, effsize_rules="cohen1988", ...) {
     outcome,
     " (formula = ",
     stringr::str_squish(paste0(format(eval(fit@call$formula)), collapse = "")),
-    ") successfully converged",
-    " and explained ",
+    ") successfully converged and explained ",
     format_digit(R2c * 100, 2),
     "% of the variance of the endogen (the conditional R2). ",
     "The variance explained by the fixed effects was of ",
@@ -196,7 +202,7 @@ analyze.lmerModLmerTest <- function(x, CI=95, effsize_rules="cohen1988", ...) {
   # -------------
   plot <- "Not available yet"
 
-  output <- list(text = text, plot = plot, summary = fitsum, values = values)
+  output <- list(text = text, plot = plot, summary = summary, values = values)
 
   class(output) <- c("psychobject", "list")
   return(output)