+++ tests/testthat/test-analyze.glmerMod.R
@@ -13,4 +13,8 @@ test_that("If it works.", {
   # test summary
   summa <- summary(model, round = 2)
   testthat::expect_equal(nrow(summa), 2)
+
+  # GLM
+  fit <- lme4::glmer(Sex ~ Adjusting + (1 | Salary), data = psycho::affective, family = "binomial")
+  testthat::expect_warning(analyze(fit))
 })