+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Breaking changes
 - Removed Mean and SD from summary in `bayes_cor.test`
 ### New functions / parameters
+- Added `interpret_odds` and logistic regression effect size interpretation
 - Added `omega_sq` and `interpret_omega_sq`
 ### Changes
 - Enhance print for `analyze.aov`